Fix the EndlessListComponent not reliably fetching
Need to use debounceTime() instead of throttleTime() here.  The latter
is faster, because it will relay the first event, and swallow all that
happen soon after, but this means if the user hits the end key on the
keyboard, the first event will possibly be way up and not fetch
anything, but since the page then scrolls all the way to the bottom
almost instantly, all other scroll events will be eaten by the throttle
and the list will fail to fetch more content.